Show empty state when a user has no joinable channels
And also extract channel browsing tests into their own file
